Time: 2 ms (99.87%), Space: 56.4 MB (54.13%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -1,13 +1,13 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
-        double max_sum=0;
+        int max_sum=0;
         for(int i=0 ; i<k ; i++)max_sum+=nums[i];
-        double sum = max_sum;
+        int sum = max_sum;
         for(int i=k ; i<nums.length ; i++)
         {
             sum = sum+nums[i]-nums[i-k];
             max_sum = Math.max(max_sum,sum);
         }
-        return max_sum/k;
+        return ((double)max_sum/k);
     }
 }
